+++ src/main/components/physicsComponent/Collider.java
@@ -2,10 +2,11 @@ package main.components.physicsComponent;
 
 import main.editor.JImGui;
 import main.components.Component;
+import org.joml.Vector2d;
 import org.joml.Vector2f;
 
 public class Collider extends Component {
-    private Vector2f offset = new Vector2f();
+    private Vector2d offset = new Vector2d();
 
     @Override
     public void dearGui(){
@@ -16,11 +17,11 @@ public class Collider extends Component {
     public void update(float dt) {
     }
 
-    public void setOffset(Vector2f offset){
+    public void setOffset(Vector2d offset){
         this.offset = offset;
     }
 
-    public Vector2f getOffset(){
+    public Vector2d getOffset(){
         return offset;
     }
 }
